pyproject.toml
[project]
name = "omop-semantics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "gkennos", email = "georgina.kennedy@unsw.edu.au" }
]
requires-python = ">=3.12"
dependencies = [
    "linkml-runtime>=1.9.5",
    "python-dotenv>=1.2.1",
    "ruamel-yaml>=0.18.17",
    "sqlalchemy>=2.0.45",
    "typing-extensions>=4.15.0",
]

[project.scripts]
omop-semantics = "omop_semantics:main"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
